+++ common-libs/user-management/src/index.ts
@@ -1,0 +1,4 @@
+export * from './UserManagementService'
+export * from './validateUsername'
+export * from './normalizeUsername'
+export * from './CognitoIdentityService'