+++ common-libs/user-management/test/factory/cognitoUserTokens.ts
@@ -1,0 +1,6 @@
+export const cognitoUserTokens = {
+  idToken: 'dummy_token',
+  accessToken: 'dummy_token',
+  refreshToken: 'dummy_token',
+  expiresIn: 1591890477,
+}