+++ common-libs/user-management/test/integration/index.test.ts
@@ -1,0 +1,7 @@
+import { expect } from 'chai'
+
+describe('[stub]', () => {
+  it('should pass', () => {
+    expect(true).to.be.true
+  })
+})