+++ sdk/core/test/integration/NetworkMember.test.ts
@@ -5,7 +5,6 @@ import { decode as jwtDecode } from 'jsonwebtoken'
 import { Affinity, KeysService } from '@affinidi/common'
 import { buildVCV1Unsigned, buildVCV1Skeleton } from '@affinidi/vc-common'
 import { VCSPhonePersonV1, getVCPhonePersonV1Context } from '@affinidi/vc-data'
-import UserManagementService from '../../src/services/UserManagementService'
 import { AffinidiWallet, checkIsWallet } from '../helpers/AffinidiWallet'
 
 import {
@@ -20,6 +19,7 @@ import { generateCredentials } from '../helpers/generateCredentials'
 import { SignedCredential, SdkOptions } from '../../src/dto'
 import { testPlatformTools } from '../helpers/testPlatformTools'
 import { RegistryApiService } from '@affinidi/internal-api-clients'
+import { createUserManagementService } from '../../src/shared/createUserManagementService'
 
 const {
   PASSWORD,
@@ -1017,7 +1017,7 @@ describe('CommonNetworkMember', () => {
     //       and `options` has only accessApiKey and env
     //       This is important for testing against different environments
     const fullOptions = getAllOptionsForEnvironment()
-    const userManagementService = new UserManagementService(fullOptions)
+    const userManagementService = createUserManagementService({ ...fullOptions, basicOptions: fullOptions })
 
     const { idToken } = await userManagementService.logInWithPassword(cognitoUsername, cognitoPassword)